Library rearranged to allow multiple DyIO connections.
